requirements.txt
mecab-python3
requests
beautifulsoup4
lxml
sklearn
numpy
scipy
